+++ src/Ext/Bullet/Trajectories/BombardTrajectory.h
@@ -7,6 +7,14 @@ class BombardTrajectoryType final : public PhobosTrajectoryType
 public:
 	BombardTrajectoryType() : PhobosTrajectoryType(TrajectoryFlag::Bombard)
 		, Height { 0.0 }
+		, FallPercent { 1.0 }
+		, FallPercentShift { 0.0 }
+		, FallScatterRange { Leptons(0) }
+		, FallSpeed { 0.0 }
+		, TargetSnapDistance { Leptons(128) }
+		, FreeFallOnTarget { true }
+		, NoLaunch { false }
+		, TurningPointAnim {}
 	{}
 
 	virtual bool Load(PhobosStreamReader& Stm, bool RegisterForChange) override;
@@ -15,7 +23,15 @@ public:
 
 	virtual void Read(CCINIClass* const pINI, const char* pSection) override;
 
-	double Height;
+	Valueable<double> Height;
+	Valueable<double> FallPercent;
+	Valueable<double> FallPercentShift;
+	Valueable<Leptons> FallScatterRange;
+	Valueable<double> FallSpeed;
+	Valueable<Leptons> TargetSnapDistance;
+	Valueable<bool> FreeFallOnTarget;
+	Valueable<bool> NoLaunch;
+	Valueable<AnimTypeClass*> TurningPointAnim;
 };
 
 class BombardTrajectory final : public PhobosTrajectory
@@ -24,11 +40,27 @@ public:
 	BombardTrajectory() : PhobosTrajectory(TrajectoryFlag::Bombard)
 		, IsFalling { false }
 		, Height { 0.0 }
+		, RemainingDistance { 1 }
+		, FallPercent { 1.0 }
+		, FallPercentShift { 0.0 }
+		, FallScatterRange { Leptons(0) }
+		, FallSpeed { 0.0 }
+		, TargetSnapDistance { Leptons(128) }
+		, FreeFallOnTarget { true }
+		, NoLaunch { false }
 	{}
 
 	BombardTrajectory(PhobosTrajectoryType const* pType) : PhobosTrajectory(TrajectoryFlag::Bombard)
 		, IsFalling { false }
 		, Height { 0.0 }
+		, RemainingDistance { 1 }
+		, FallPercent { 1.0 }
+		, FallPercentShift { 0.0 }
+		, FallScatterRange { Leptons(0) }
+		, FallSpeed { 0.0 }
+		, TargetSnapDistance { Leptons(128) }
+		, FreeFallOnTarget { true }
+		, NoLaunch { false }
 	{}
 
 	virtual bool Load(PhobosStreamReader& Stm, bool RegisterForChange) override;
@@ -36,11 +68,22 @@ public:
 
 	virtual void OnUnlimbo(BulletClass* pBullet, CoordStruct* pCoord, BulletVelocity* pVelocity) override;
 	virtual bool OnAI(BulletClass* pBullet) override;
-	virtual void OnAIPreDetonate(BulletClass* pBullet) override {};
+	virtual void OnAIPreDetonate(BulletClass* pBullet) override;
 	virtual void OnAIVelocity(BulletClass* pBullet, BulletVelocity* pSpeed, BulletVelocity* pPosition) override;
 	virtual TrajectoryCheckReturnType OnAITargetCoordCheck(BulletClass* pBullet) override;
 	virtual TrajectoryCheckReturnType OnAITechnoCheck(BulletClass* pBullet, TechnoClass* pTechno) override;
 
 	bool IsFalling;
 	double Height;
+	int RemainingDistance;
+	double FallPercent;
+	double FallPercentShift;
+	Leptons FallScatterRange;
+	double FallSpeed;
+	Leptons TargetSnapDistance;
+	bool FreeFallOnTarget;
+	bool NoLaunch;
+
+private:
+	bool BulletDetonatePreCheck(BulletClass* pBullet, HouseClass* pOwner, double StraightSpeed);
 };